+++ jest.browser.config.js
@@ -1,6 +1,8 @@
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
 const nodeCrypto = require('crypto');
+const browserCrypto = require('crypto-browserify');
+process.env.RUNTIME_ENV = 'browser';
 module.exports = {
   // All imported modules in your tests should be mocked automatically
   // automock: false,
@@ -70,7 +72,8 @@ module.exports = {
   // A set of global variables that need to be available in all test environments
   globals: {
     crypto: {
-      getRandomValues: function(buffer) { return nodeCrypto.randomFillSync(buffer);}
+      getRandomValues: function(buffer) { return nodeCrypto.randomFillSync(buffer);},
+      ...browserCrypto
     }
   },
 
@@ -90,7 +93,9 @@ module.exports = {
   // ],
 
   // A map from regular expressions to module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
+  moduleNameMapper: {
+    '^scrypt.js$': 'scrypt.js/js.js'
+  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],