+++ jest.node.config.js
@@ -1,6 +1,8 @@
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
 
+process.env.RUNTIME_ENV = 'node';
+
 module.exports = {
   // All imported modules in your tests should be mocked automatically
   // automock: false,
@@ -77,7 +79,9 @@ module.exports = {
   // ],
 
   // A map from regular expressions to module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
+  moduleNameMapper: {
+    '^scrypt.js$': 'scrypt.js/node.js'
+  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],