+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.2.10](https://github.com/AElfProject/aelf-sdk.js/compare/v3.2.9...v3.2.10) (2019-07-17)
+
+
+### Features
+
+* 🎸 rewrite keyStore generate and resolve ([6ed2b5d](https://github.com/AElfProject/aelf-sdk.js/commit/6ed2b5d))
+
+
+
+### [3.2.9](https://github.com/AElfProject/aelf-sdk.js/compare/v3.2.8...v3.2.9) (2019-07-16)
+
+
+### Bug Fixes
+
+* 🐛 fix multi-request share the same xhr instance ([b7a20eb](https://github.com/AElfProject/aelf-sdk.js/commit/b7a20eb))
+
+
+
+### [3.2.8](https://github.com/AElfProject/aelf-sdk.js/compare/v3.2.7...v3.2.8) (2019-07-15)
+
+
+### Bug Fixes
+
+* 🐛 fix toString error ([8f0da78](https://github.com/AElfProject/aelf-sdk.js/commit/8f0da78))
+
+
+### Features
+
+* 🎸 add compile env variables for import correct xhr lib ([a767bd6](https://github.com/AElfProject/aelf-sdk.js/commit/a767bd6))
+
+
+
 ### [3.2.7](https://github.com/AElfProject/aelf-sdk.js/compare/v3.2.6...v3.2.7) (2019-07-13)